requirements.txt
aiogram
python-dotenv
pandas
aiohttp
bs4
requests
joblib
scikit-learn
xgboost